extract js semantic tokens
