+++ extensions/html-language-features/server/src/modes/languageModes.ts
@@ -17,7 +17,8 @@ import { getDocumentRegions, HTMLDocumentRegions } from './embeddedSupport';
 import { getHTMLMode } from './htmlMode';
 import { getJavaScriptMode } from './javascriptMode';
 
-export { ColorInformation, ColorPresentation, Color };
+export * from 'vscode-html-languageservice';
+export { WorkspaceFolder } from 'vscode-languageserver';
 
 export interface Settings {
 	css?: any;
